Remove client.py and move code to __init__.py
